Refactor enums to a their own directory
