+++ Modules/Dnd5e/app/Enums/DamageType.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Modules\Dnd5e\Models;
+namespace Modules\Dnd5e\Enums;
 
 enum DamageType: string
 {
