+++ Modules/Expanse/app/Enums/CrewCompetence.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Modules\Expanse\Models;
+namespace Modules\Expanse\Enums;
 
 enum CrewCompetence: string
 {
