+++ Modules/Shadowrun5e/app/Enums/VehicleModificationSlotType.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Modules\Shadowrun5e\Models;
+namespace Modules\Shadowrun5e\Enums;
 
 enum VehicleModificationSlotType: string
 {
