+++ Modules/Shadowrun5e/app/Models/Lifestyle.php
@@ -111,7 +111,7 @@ class Lifestyle implements Stringable
      */
     public static function all(): array
     {
-        $filename = config('app.data_path.shadowrun5e') . 'lifestyles.php';
+        $filename = config('shadowrun5e.data_path') . 'lifestyles.php';
         self::$lifestyles ??= require $filename;
 
         $lifestyles = [];
