+++ Modules/Transformers/app/Enums/Action.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Modules\Transformers\Models;
+namespace Modules\Transformers\Enums;
 
 enum Action
 {
