+++ Modules/Transformers/app/Http/Requests/ProgrammingRequest.php
@@ -7,7 +7,7 @@ namespace Modules\Transformers\Http\Requests;
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Validation\Rule;
 use Illuminate\Validation\Rules\In;
-use Modules\Transformers\Models\Programming;
+use Modules\Transformers\Enums\Programming;
 
 class ProgrammingRequest extends FormRequest
 {
