+++ Modules/Transformers/tests/Feature/Enums/ProgrammingTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Modules\Transformers\Tests\Feature\Models;
+namespace Modules\Transformers\Tests\Feature\Enums;
 
-use Modules\Transformers\Models\Action;
-use Modules\Transformers\Models\Programming;
+use Modules\Transformers\Enums\Action;
+use Modules\Transformers\Enums\Programming;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Group;
 use PHPUnit\Framework\Attributes\Small;
