+++ app/Console/Commands/ImportChummerData.php
@@ -8,9 +8,9 @@ use GitElephant\Repository;
 use Illuminate\Console\Command;
 use Illuminate\Contracts\Console\Isolatable;
 use Illuminate\Support\Str;
+use Modules\Shadowrun5e\Enums\VehicleModificationSlotType;
+use Modules\Shadowrun5e\Enums\VehicleModificationType;
 use Modules\Shadowrun5e\Models\ForceTrait;
-use Modules\Shadowrun5e\Models\VehicleModificationSlotType;
-use Modules\Shadowrun5e\Models\VehicleModificationType;
 use RuntimeException;
 use SimpleXMLElement;
 use Throwable;
